create backingstore_factory_session
