+++ tests/conftest.py
@@ -33,7 +33,9 @@ from ocs_ci.ocs.exceptions import TimeoutExpiredError, CephHealthException, Reso
 from ocs_ci.ocs.node import get_node_objs, schedule_nodes
 from ocs_ci.ocs.ocp import OCP
 from ocs_ci.ocs.utils import setup_ceph_toolbox
-from ocs_ci.ocs.resources.backingstore import BackingStore
+from ocs_ci.ocs.resources.backingstore import (
+    backingstore_factory as backingstore_factory_implementation
+)
 from ocs_ci.ocs.resources.cloud_manager import CloudManager
 from ocs_ci.ocs.node import check_nodes_specs
 from ocs_ci.ocs.resources.mcg import MCG
@@ -66,12 +68,7 @@ from ocs_ci.utility.utils import (
 )
 from tests import helpers
 from tests.helpers import create_unique_resource_name
-from ocs_ci.ocs.bucket_utils import (
-    oc_create_aws_backingstore, oc_create_google_backingstore, oc_create_azure_backingstore,
-    oc_create_s3comp_backingstore, oc_create_pv_backingstore, cli_create_aws_backingstore,
-    cli_create_google_backingstore, cli_create_azure_backingstore, cli_create_s3comp_backingstore,
-    cli_create_pv_backingstore, get_rgw_restart_counts
-)
+from ocs_ci.ocs.bucket_utils import get_rgw_restart_counts
 from ocs_ci.ocs.pgsql import Postgresql
 from ocs_ci.ocs.resources.rgw import RGW
 from ocs_ci.ocs.jenkins import Jenkins
@@ -2055,97 +2052,38 @@ def backingstore_factory(request, cld_mgr, cloud_uls_factory):
             cloud_uls_factory: Factory for underlying storage creation
             cld_mgr (CloudManager): Cloud Manager object containing all connections to clouds
 
+        Returns:
+            func: Factory method - each call to this function creates
+                a backingstore
+
     """
-    created_backingstores = []
+    return backingstore_factory_implementation(
+        request,
+        cld_mgr,
+        cloud_uls_factory
+    )
 
-    cmdMap = {
-        'oc': {
-            'aws': oc_create_aws_backingstore,
-            'google': oc_create_google_backingstore,
-            'azure': oc_create_azure_backingstore,
-            'ibmcos': oc_create_s3comp_backingstore,
-            'pv': oc_create_pv_backingstore
-        },
-        'cli': {
-            'aws': cli_create_aws_backingstore,
-            'google': cli_create_google_backingstore,
-            'azure': cli_create_azure_backingstore,
-            'ibmcos': cli_create_s3comp_backingstore,
-            'pv': cli_create_pv_backingstore
-        }
-    }
 
-    def _create_backingstore(method, uls_dict):
-        """
-        Tracks creation and cleanup of all the backing stores that were created in the scope
+@pytest.fixture(scope='session')
+def backingstore_factory_session(request, cld_mgr, cloud_uls_factory):
+    """
+        Create a Backing Store factory.
+        Calling this fixture creates a new Backing Store(s).
 
         Args:
-            method (str): String for selecting method of backing store creation (CLI/OC)
-            uls_dict (dict): Dictionary containing storage provider as key and a list of tuples
-            as value.
-            Cloud backing stores form - 'CloudName': [(amount, region), (amount, region)]
-            i.e. - 'aws': [(3, us-west-1),(2, eu-west-2)]
-            PV form - 'pv': [(amount, size_in_gb, storageclass), ...]
-            i.e. - 'pv': [(3, 32, ocs-storagecluster-ceph-rbd),(2, 100, ocs-storagecluster-ceph-rbd)]
+            cloud_uls_factory: Factory for underlying storage creation
+            cld_mgr (CloudManager): Cloud Manager object containing all connections to clouds
 
         Returns:
-            list: A list of backingstore names.
+            func: Factory method - each call to this function creates
+                a backingstore
 
-        """
-        if method.lower() not in cmdMap:
-            raise RuntimeError(
-                f'Invalid method type received: {method}. '
-                f'available types: {", ".join(cmdMap.keys())}'
-            )
-        for cloud, uls_lst in uls_dict.items():
-            for uls_tup in uls_lst:
-                # Todo: Replace multiple .append calls, create names in advance, according to amountoc
-                if cloud.lower() not in cmdMap[method.lower()]:
-                    raise RuntimeError(
-                        f'Invalid cloud type received: {cloud}. '
-                        f'available types: {", ".join(cmdMap[method.lower()].keys())}'
-                    )
-                if cloud == 'pv':
-                    vol_num, size, storage_class = uls_tup
-                    backingstore_name = create_unique_resource_name(
-                        resource_description='backingstore', resource_type=cloud.lower()
-                    )
-                    # removing characters from name (pod name length bellow 64 characters issue)
-                    backingstore_name = backingstore_name[:-16]
-                    created_backingstores.append(backingstore_name)
-                    cmdMap[method.lower()][cloud.lower()](
-                        backingstore_name, vol_num, size, storage_class
-                    )
-                else:
-                    region = uls_tup[1]
-                    # Todo: Verify that the given cloud has an initialized client
-                    uls_dict = cloud_uls_factory({cloud: [uls_tup]})
-                    for uls_name in uls_dict[cloud.lower()]:
-                        backingstore_name = create_unique_resource_name(
-                            resource_description='backingstore', resource_type=cloud.lower()
-                        )
-                        # removing characters from name (pod name length bellow 64 characters issue)
-                        backingstore_name = backingstore_name[:-16]
-                        created_backingstores.append(
-                            BackingStore(
-                                name=backingstore_name,
-                                uls_name=uls_name
-                            )
-                        )
-                        cmdMap[method.lower()][cloud.lower()](
-                            cld_mgr, backingstore_name, uls_name, region
-                        )
-                        # Todo: Raise an exception in case the BS wasn't created
-
-        return created_backingstores
-
-    def backingstore_cleanup():
-        for backingstore in created_backingstores:
-            backingstore.delete()
-
-    request.addfinalizer(backingstore_cleanup)
-
-    return _create_backingstore
+    """
+    return backingstore_factory_implementation(
+        request,
+        cld_mgr,
+        cloud_uls_factory
+    )
 
 
 @pytest.fixture()
@@ -2190,13 +2128,13 @@ def multiregion_mirror_setup(mcg_obj, multiregion_resources, backingstore_factor
 def multiregion_mirror_setup_session(
     mcg_obj_session,
     multiregion_resources_session,
-    backingstore_factory,
+    backingstore_factory_session,
     bucket_factory_session
 ):
     return multiregion_mirror_setup_fixture(
         mcg_obj_session,
         multiregion_resources_session,
-        backingstore_factory,
+        backingstore_factory_session,
         bucket_factory_session
     )
